+++ tests/step_defs/test_grype.py
@@ -1,21 +1,37 @@
 # coding=utf-8
 """The Grype Container feature tests."""
-
+import importlib
+import os
 import testinfra_bdd
 
-from pytest_bdd import scenario
+from pytest_bdd import (
+    scenarios,
+    given,
+    then,
+    parsers
+)
 
 # Ensure that the PyTest fixtures provided in testinfra-bdd are available to
 # your test suite.
 pytest_plugins = testinfra_bdd.PYTEST_MODULES
+scenarios('../features/grype.feature')
+grype_parser = importlib.import_module('parse-grype-json')
 
 
-@scenario('../features/grype.feature',
-          'Assert the Container is Built as Expected')
-def test_assert_the_container_is_built_as_expected():
-    """Assert the Container is Build as Expected."""
+@given(parsers.parse('VULNERABILITIES_ALLOWED_LIST="{allowed_list}"'), target_fixture='vulnerabilities_allowed_list')
+def vulnerabilities_allowed_list(allowed_list):
+    """VULNERABILITIES_ALLOWED_LIST="<allowed_list>"."""
+    return allowed_list
 
 
-@scenario('../features/grype.feature', 'Test Script Output and Exit Code')
-def test_test_script_output_and_exit_code():
-    """Test Script Output and Exit Code."""
+@given('VULNERABILITIES_ALLOWED_LIST=""', target_fixture='vulnerabilities_allowed_list')
+def vulnerabilities_allowed_list_is_blank():
+    """VULNERABILITIES_ALLOWED_LIST="<allowed_list>"."""
+    return ''
+
+
+@then(parsers.parse('main return code is {expected_exit_code:d}'))
+def the_command_return_code_is_expected_exit_code(expected_exit_code, vulnerabilities_allowed_list):
+    """the command return code is <expected_exit_code>."""
+    os.environ['VULNERABILITIES_ALLOWED_LIST'] = vulnerabilities_allowed_list
+    assert grype_parser.main('tests/resources/grype.json') == expected_exit_code